So i finally managed to move the elements to a new function...
lowkey this needs another f*ckin rewrite im crying
